Fixed Merge Conflicts and Got DB working

+++ app/src/main/java/sep/myapplication/Timer.java
@@ -28,7 +28,7 @@ class Timer {
     }
 
     public String toString(){
-        int Seconds, Minutes, MilliSeconds;
+        long Seconds, Minutes, MilliSeconds;
 
         MilliSeconds = (int)run();
